+++ app/[domain]/(default)/explore/page.tsx
@@ -1,9 +1,3 @@
 import { ExploreContent } from '$lib/components/features/explore/Explore';
-import { notFound } from 'next/navigation';
 
-function Page() {
-  if (process.env.NEXT_PUBLIC_APP_ENV === 'production') return notFound();
-
-  return <ExploreContent />;
-}
-export default Page;
+export default ExploreContent;