+++ lib/trpc/index.ts
@@ -4,6 +4,7 @@ import { match } from 'ts-pattern';
 import net from 'net';
 import { lookup } from 'dns/promises';
 import { isPrivateIP } from 'range_check';
+import { PinataSDK } from 'pinata';
 
 import { getMintNftData } from '$lib/services/lemonhead';
 import { calculateLookHash, Filter, getFinalTraits, validateTraits, type Trait } from '$lib/services/lemonhead/core';
@@ -17,7 +18,6 @@ import { getMintVinylNationPassportImage } from '$lib/services/passports/vinyl-n
 import { getMintDripNationPassportImage } from '$lib/services/passports/drip-nation';
 import { getMintFestivalNationPassportImage } from '$lib/services/passports/festival-nation';
 import { request } from '$lib/services/nft/admin';
-import { pinata } from '$lib/utils/pinata';
 
 export const appRouter = router({
   ping: publicProcedure.query(async () => {
@@ -178,7 +178,7 @@ export const appRouter = router({
       const { username } = input;
       const metadata = {
         name: username,
-        description: 'Lemonade Username',
+        description: 'Lemonade usernames powered by MegaETH! #makelemonade',
         attributes: [
           {
             trait_type: 'Created Date',
@@ -196,6 +196,10 @@ export const appRouter = router({
       const metadataBlob = new Blob([JSON.stringify(metadata)], { type: 'application/json' });
       const metadataFile = new File([metadataBlob], 'metadata.json', { type: 'application/json' });
 
+      const pinata = new PinataSDK({
+        pinataJwt: process.env.USERNAME_PINATA_JWT,
+      });
+      
       const { cid: metadataCid } = await pinata.upload.public.file(metadataFile);
       const tokenUri = `ipfs://${metadataCid}`;